Add missing generator information for HTML and manual pages
The CHM/PS/PDF/Info files already have generator information.
The paper backend seems to pass data to the PS/PDF backends.
The WinHelp files seem difficult to add generator information to.
The text files don't have a way to add hidden by default metadata.

The HTML gets a <meta> tag with generator/content attribute.
The manual pages get a comment at the start of the document.

This enables folks interacting with the generated files to know
that they are not manually edited files and thus the Halibut
source files should be edited instead.